01_az_rag_poc/requirements.txt
# Azure SDK
azure-identity==1.15.0
azure-storage-blob==12.19.0
azure-search-documents==11.4.0

# Azure OpenAI (use openai SDK with Azure endpoint)
openai==1.12.0

# Document Processing
pypdf==4.0.1
tiktoken==0.6.0

# Web UI
streamlit==1.31.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
loguru==0.7.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
reportlab==4.1.0  # For generating test PDFs
